+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `FetchTrophy()` for getting informations about a trophy.
+- `ListTrophies()` to get a list of trophies.
 
 ## 0.2.0
 ### Added